+++ src/app/utils/utils.ts
@@ -88,10 +88,7 @@ export function hasValue(v: any): boolean {
 }
 
 export function titlePlural(s: string): string {
-  if ((s == 'Art' || s == 'Camps')) {
-    return s;
-  }
-  return s + 's';
+  return s;
 }
 
 export function delay(ms: number): Promise<void> {